# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Тропа памяти: Последний бой — API",
    "version": "0.1.0",
    "description": "Контракт, который ожидает фронтенд. Источник правды — frontend/src/contract (zod). Сгенерировано: npm run contract."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Тот же домен, что и фронт (прокси)"
    }
  ],
  "paths": {
...
    "/requests": {
      "get": {
        "operationId": "listRequests",
        "summary": "Заявки на набор волонтёров, новые сверху",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VolunteerRequest"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        }
      },
      "post": {
        "operationId": "createRequest",
        "summary": "Создать заявку (командир)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolunteerRequest"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewVolunteerRequest"
              }
            }
          }
        }
      }
    },
    "/requests/{id}/join": {
      "post": {
        "operationId": "joinRequest",
        "summary": "Записаться в заявку отряда (после окна с условиями; до 18+ — с согласием родителя)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolunteerRequest"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        }
      }
    },
    "/fundraisers": {
      "get": {
        "operationId": "listFundraisers",
        "summary": "Целевые сборы",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fundraiser"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        }
      }
    },
    "/donations": {
      "post": {
        "operationId": "donate",
        "summary": "Пожертвование (только тестовый режим ЮKassa/CloudPayments)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonationResult"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Donation"
              }
            }
          }
        }
      }
    },
...
    "/trips": {
      "get": {
        "operationId": "listTrips",
        "summary": "Выезды «Выходные с поисковиком» по дате",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        }
      }
    },
    "/trips/{id}": {
      "get": {
        "operationId": "getTrip",
        "summary": "Выезд",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/trips/{id}/register": {
      "post": {
        "operationId": "registerTrip",
        "summary": "Записаться на выезд (после окна с условиями; до 18+ — с согласием родителя)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        }
      }
    },
    "/group-applications": {
      "get": {
        "operationId": "listGroupApplications",
        "summary": "Коллективные заявки на выезды (командиру — заявки на выезды его отряда)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupApplication"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        }
      },
      "post": {
        "operationId": "createGroupApplication",
        "summary": "Подать коллективную заявку на выезд (школа, клуб, семейная группа)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupApplication"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewGroupApplication"
              }
            }
          }
        }
      }
    },
    "/group-applications/{id}": {
      "patch": {
        "operationId": "decideGroupApplication",
        "summary": "Решение командира: подтвердить или попросить уточнить",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupApplication"
                }
              }
            }
          },
          "4XX": {
            "description": "Ошибка запроса: { \"message\": string }"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupApplicationDecision"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VolunteerRole": {
        "type": "string",
        "enum": [
          "digger",
          "prober",
          "cook",
          "driver",
          "medic",
          "any"
        ],
        "description": "Кого ищет отряд"
      },
      "VolunteerRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "teamId": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "date": {
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
            "description": "Дата YYYY-MM-DD"
          },
          "place": {
            "type": "string",
            "minLength": 1
          },
          "roles": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "$ref": "#/components/schemas/VolunteerRole"
                },
                "count": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "role",
                "count"
              ],
              "additionalProperties": false
            }
          },
          "joined": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "minAge": {
            "description": "Минимальный возраст волонтёра, лет: в ленте «16+». Младше — только с родителями",
            "type": "integer",
            "minimum": 10,
            "maximum": 99
          },
          "startsAt": {
            "description": "Начало — время сбора участников",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "endsAt": {
            "description": "Окончание",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "meetingPoint": {
            "description": "Место сбора участников",
            "type": "string",
            "minLength": 1
          },
          "bring": {
            "description": "Что взять с собой",
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "fundraiserId": {
            "type": "string",
            "minLength": 1
          },
          "lat": {
            "description": "Где нужны люди — для карты потребностей",
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lon": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "description": "Дата и время ISO 8601, UTC"
          },
          "demo": {
            "type": "boolean",
            "description": "true — демо-данные, вымышлены или не проверены"
          }
        },
        "required": [
          "id",
          "teamId",
          "title",
          "date",
          "place",
          "roles",
          "joined",
          "createdAt",
          "demo"
        ],
        "additionalProperties": false,
        "description": "Заявка отряда на набор волонтёров"
      },
      "NewVolunteerRequest": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "date": {
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
            "description": "Дата YYYY-MM-DD"
          },
          "place": {
            "type": "string",
            "minLength": 1
          },
          "roles": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "$ref": "#/components/schemas/VolunteerRole"
                },
                "count": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "role",
                "count"
              ],
              "additionalProperties": false
            }
          },
          "minAge": {
            "description": "Минимальный возраст волонтёра, лет: в ленте «16+». Младше — только с родителями",
            "type": "integer",
            "minimum": 10,
            "maximum": 99
          }
        },
        "required": [
          "teamId",
          "title",
          "date",
          "place",
          "roles"
        ],
        "additionalProperties": false,
        "description": "Создание заявки командиром"
      },
      "FundraiserPurpose": {
        "type": "string",
        "enum": [
          "fuel",
          "raise_fighter",
          "equip"
        ],
        "description": "fuel — «Пожертвовать на бензин», raise_fighter — «Поднять бойца», equip — «Экипировать отряд»"
      },
      "Fundraiser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "teamId": {
            "type": "string",
            "minLength": 1
          },
          "purpose": {
            "$ref": "#/components/schemas/FundraiserPurpose"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "goalRub": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Сумма в рублях"
          },
          "collectedRub": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Сумма в рублях"
          },
          "lat": {
            "description": "Куда пойдут деньги — для карты потребностей",
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lon": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "createdAt": {
            "description": "Когда опубликовано: лента «Мероприятия» идёт по этой дате, новые сверху",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "demo": {
            "type": "boolean",
            "description": "true — демо-данные, вымышлены или не проверены"
          }
        },
        "required": [
          "id",
          "teamId",
          "purpose",
          "title",
          "goalRub",
          "collectedRub",
          "demo"
        ],
        "additionalProperties": false,
        "description": "Целевой сбор отряда"
      },
      "Donation": {
        "type": "object",
        "properties": {
          "fundraiserId": {
            "type": "string",
            "minLength": 1
          },
          "amountRub": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Сумма в рублях"
          }
        },
        "required": [
          "fundraiserId",
          "amountRub"
        ],
        "additionalProperties": false,
        "description": "Пожертвование. Только тестовый режим платёжного провайдера"
      },
      "DonationResult": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "test_succeeded"
            ]
          },
          "fundraiser": {
            "$ref": "#/components/schemas/Fundraiser"
          }
        },
        "required": [
          "paymentId",
          "status",
          "fundraiser"
        ],
        "additionalProperties": false,
        "description": "Результат тестового платежа"
      },
...
      "ChecklistItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "label": {
            "type": "string",
            "minLength": 1
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "additionalProperties": false,
        "description": "Пункт чек-листа новичка"
      },
      "Trip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "teamId": {
            "type": "string",
            "minLength": 1
          },
          "date": {
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
            "description": "Дата YYYY-MM-DD"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "place": {
            "type": "string",
            "minLength": 1
          },
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lon": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "spotsTotal": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991
          },
          "spotsTaken": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "minAge": {
            "description": "Минимальный возраст волонтёра, лет: в ленте «16+». Младше — только с родителями",
            "type": "integer",
            "minimum": 10,
            "maximum": 99
          },
          "startsAt": {
            "description": "Начало — время сбора участников",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "endsAt": {
            "description": "Окончание",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "meetingPoint": {
            "description": "Место сбора участников",
            "type": "string",
            "minLength": 1
          },
          "checklist": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "description": "Чек-лист новичка — он же «что взять с собой»"
          },
          "createdAt": {
            "description": "Когда опубликовано: лента «Мероприятия» идёт по этой дате, новые сверху",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "demo": {
            "type": "boolean",
            "description": "true — демо-данные, вымышлены или не проверены"
          }
        },
        "required": [
          "id",
          "teamId",
          "date",
          "title",
          "place",
          "lat",
          "lon",
          "spotsTotal",
          "spotsTaken",
          "checklist",
          "demo"
        ],
        "additionalProperties": false,
        "description": "Выезд «Выходные с поисковиком»"
      },
      "ParentConsent": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 5,
            "maxLength": 120,
            "description": "ФИО родителя или законного представителя"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+7\\d{10}$",
            "description": "Телефон родителя для связи, +7XXXXXXXXXX"
          },
          "agreedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "description": "Когда родитель дал согласие"
          }
        },
        "required": [
          "fullName",
          "phone",
          "agreedAt"
        ],
        "additionalProperties": false,
        "description": "Согласие родителя или законного представителя на участие несовершеннолетнего (или без подтверждённых 18+)"
      },
      "SignupRequest": {
        "type": "object",
        "properties": {
          "termsAccepted": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Человек прочитал условия: время, место сбора, возраст"
          },
          "adultVerified": {
            "type": "boolean",
            "description": "Возраст 18+ подтверждён в профиле"
          },
          "age": {
            "description": "Возраст участника, если известен из профиля: сервер сверяет его с minAge",
            "type": "integer",
            "minimum": 1,
            "maximum": 120
          },
          "parentConsent": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ParentConsent"
              }
            ]
          }
        },
        "required": [
          "termsAccepted",
          "adultVerified"
        ],
        "additionalProperties": false,
        "description": "Запись на заявку отряда или выезд после окна с условиями. Без подтверждённого возраста 18+ — только с согласием родителя"
      },
...
      "GroupApplication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "tripId": {
            "type": "string",
            "minLength": 1
          },
          "organization": {
            "type": "string",
            "minLength": 1,
            "description": "Школа, клуб или название группы"
          },
          "contactName": {
            "type": "string",
            "minLength": 2
          },
          "contact": {
            "type": "string",
            "minLength": 3,
            "description": "Телефон или электронная почта ответственного"
          },
          "peopleCount": {
            "type": "integer",
            "minimum": 2,
            "maximum": 100
          },
          "comment": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/GroupApplicationStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "description": "Дата и время ISO 8601, UTC"
          },
          "demo": {
            "type": "boolean",
            "description": "true — демо-данные, вымышлены или не проверены"
          }
        },
        "required": [
          "id",
          "tripId",
          "organization",
          "contactName",
          "contact",
          "peopleCount",
          "comment",
          "status",
          "createdAt",
          "demo"
        ],
        "additionalProperties": false,
        "description": "Коллективная заявка на выезд: школа, клуб или семейная группа. Контакт видит только командир отряда"
      },
      "NewGroupApplication": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "minLength": 1
          },
          "organization": {
            "type": "string",
            "minLength": 1,
            "description": "Школа, клуб или название группы"
          },
          "contactName": {
            "type": "string",
            "minLength": 2
          },
          "contact": {
            "type": "string",
            "minLength": 3,
            "description": "Телефон или электронная почта ответственного"
          },
          "peopleCount": {
            "type": "integer",
            "minimum": 2,
            "maximum": 100
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "tripId",
          "organization",
          "contactName",
          "contact",
          "peopleCount",
          "comment"
        ],
        "additionalProperties": false,
        "description": "Подать коллективную заявку. Статус на сервере всегда pending"
      },
      "GroupApplicationDecision": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "confirmed",
              "clarify"
            ]
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false,
        "description": "Решение командира по коллективной заявке"
      },
...
    }
  }
}